adjusted as go module
removed redis requirement
