+++ config.json
@@ -1,12 +1,12 @@
 {
   "logfile":  "/usr/local/var/log/mbtc.log",
-  "redis_url": "127.0.0.1:6379",
+  "cachefile":  "/usr/local/var/cache/mbtc.cache",
   "link_address_color": "#B0BEC5",
   "link_wallet_color": "#107896",
   "wallet_max_size": 5000,
   "cache_addresses": 1000,
   "cache_wallets": 5000,
-  "icon_address": "https://5haman.github.io/images/maltego/bitcoin.png",
-  "icon_wallet": "https://5haman.github.io/images/maltego/wallet.png",
-  "icon_service": "https://5haman.github.io/images/maltego/service.png"
+  "icon_address": "https://raw.githubusercontent.com/Megarushing/maltego-btc/master/assets/bitcoin.png",
+  "icon_wallet": "https://raw.githubusercontent.com/Megarushing/maltego-btc/master/assets/wallet.png",
+  "icon_service": "https://raw.githubusercontent.com/Megarushing/maltego-btc/master/assets/service.png"
 }
