+++ go.mod
@@ -1,0 +1,9 @@
+module github.com/Megarushing/maltego-btc
+
+go 1.16
+
+require (
+	github.com/glennzw/maltegogo v0.0.0-20160224165212-ca4722a47849
+	github.com/patrickmn/go-cache v2.1.0+incompatible
+	gonum.org/v1/gonum v0.9.1
+)
